Refactor notification context provider structure.
