+++ new-log-viewer/src/contexts/NotificationContextProvider/index.css
@@ -1,0 +1,27 @@
+.pop-up-messages-container-snackbar {
+    right: 14px !important;
+    bottom: var(--ylv-status-bar-height) !important;
+
+    padding: 0 !important;
+
+    background: transparent !important;
+    border: none !important;
+    box-shadow: none !important;
+}
+
+.pop-up-message-box-alert {
+    padding-inline: 18px !important;
+}
+
+.pop-up-message-box-alert-layout {
+    width: 300px;
+}
+
+.pop-up-message-box-title-container {
+    display: flex;
+    align-items: center;
+}
+
+.pop-up-message-box-title-text {
+    flex-grow: 1;
+}
